Add files via upload
This project is split into three parts:
Part one deals with creating Time and Duration classes and then TestofTime as the JUnit for the two classes. This Junit involves seven tests, three for Time class and four for the Duration class.
Part two involves creating classes Item and ShoppingCart and using them to write a class called TestShoppingCart that has the
main method to allow the user to capture details of item(s) purchased and print out an invoice for the
customer. 
Part three involves firstly writing a Student class to model a student's name, composed of a first
name, middle name and last name, and secondly using a supplied test suite to drive the process.